==> part01.ipynb <==
import pandas as pd



# Load the data into Pandas DataFrames
title_basics  = pd.read_csv('/Users/lidiv/project03/title.basics.tsv.gz', sep='\t', low_memory=False)
title_ratings = pd.read_csv('/Users/lidiv/project03/title.ratings.tsv.gz', sep='\t', low_memory=False)
title_akas = pd.read_csv('/Users/lidiv/project03/title.akas.tsv.gz', sep='\t', low_memory=False)



# Print info for each dataframe
print("title_basics info:")
print(title_basics.info())

print("title_ratings info:")
print(title_ratings.info())


print(title_basics.columns)
print(title_ratings.columns)
print(title_akas.columns)

# Filter movies released in the United States
title_akas = title_akas[title_akas['region'] == 'US']

# Convert startYear and endYear columns to numeric
title_basics['startYear'] = pd.to_numeric(title_basics['startYear'], errors='coerce')
title_basics['endYear'] = pd.to_numeric(title_basics['endYear'], errors='coerce')


# Filter full-length movies
title_basics = title_basics[title_basics['titleType'] == 'movie']

# Exclude movies from the Documentary genre
title_basics = title_basics[~title_basics['genres'].str.contains('Documentary', na=False)]

# Filter movies released between 2000 and 2021
title_basics = title_basics[(title_basics['startYear'] >= 2000) & (title_basics['startYear'] <= 2021)]

# Remove movies with missing values for genre or runtime
title_basics = title_basics.dropna(subset=['genres', 'runtimeMinutes'])

# Filter title_akas based on filtered title_basics
title_akas = title_akas[title_akas['titleId'].isin(title_basics['tconst'])]

# Filter title_ratings based on filtered title_basics
title_ratings = title_ratings[title_ratings['tconst'].isin(title_basics['tconst'])]

# Filter title_akas based on filtered title_basics
title_akas = title_akas[title_akas['titleId'].isin(title_basics['tconst'])]

# Save filtered dataframes to compressed CSV files
title_basics.to_csv('Data/title_basic_filtered.csv.gz', index=False, compression='gzip')
title_ratings.to_csv('Data/title_rating_filtered.csv.gz', index=False, compression='gzip')
title_akas.to_csv('Data/title_akas_filtereds.csv.gz', index=False, compression='gzip')
