==> part 03.ipynb <==
import pandas as pd
from sqlalchemy import create_engine
from sqlalchemy_utils import database_exists
import pandas as pd
from sqlalchemy import create_engine
from sqlalchemy import Text
from sqlalchemy.orm import sessionmaker
from sqlalchemy_utils import database_exists, create_database
import pymysql
pymysql.install_as_MySQLdb()
from urllib.parse import quote_plus as urlquote

from sqlalchemy import create_engine, Column, String, Float, Integer
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker

title_basics_df = pd.read_csv(r'\Users\lidiv\project03\project3\Data/title_basic_filtered.csv.gz', compression='gzip', low_memory=False)
title_ratings_df =  pd.read_csv(r'\Users\lidiv\project03\project3\Data/title_rating_filtered.csv.gz', compression='gzip', low_memory=False)
tmdb_data_df = pd.read_csv(r'\Users\lidiv\project03\project3\tmdb_results_combined.csv.gz', compression='gzip', low_memory=False)

import json
with open('/Users/lidiv/.secret/mysql.json') as f:
    login = json.load(f)
login.keys()

connection = f"mysql+mysqlconnector://{login['username']}:{urlquote(login['password'])}@localhost:3306/movie"
engine = create_engine(connection)

## Check if database exists, if not, create it
if database_exists(connection):
    print('It exists!')
else:
    create_database(connection)
    print('Database created!')

# Normalize genres
title_basics_df['genres_split'] = title_basics_df['genres'].str.split(', ')



exploded_genres = title_basics_df.explode('genres_split')
exploded_genres

!pip install sqlalchemy sqlalchemy-utils
!pip install mysql-connector-python

#unique_genres = sorted(exploded_genres['genres'].unique())

unique_genres = sorted(exploded_genres['genres_split'].unique())

titles_genres= exploded_genres[['tconst', 'genres_split']].copy()
titles_genres.head()

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## make new integer genre_id and drop string genres

title_basics_df['genre_id'] = title_basics_df['genres_split'].apply(lambda x: [genre_map[genre] for genre in x] if isinstance(x, list) else [])
title_basics_df = title_basics_df.drop(columns='genres_split')


genre_lookup = pd.DataFrame({'genre_name': genre_map.keys(),
                             'genre_ID': genre_map.values()})
genre_lookup.head()


print(title_basics_df.columns)


# Insert the DataFrame into the database
genre_lookup.to_sql('genres', engine, index=False, if_exists='replace')

title_genres = title_basics_df[['tconst', 'genre_id']]
title_genres.head()

title_genres_df= title_basics_df[['tconst', 'genre_id']].copy()
title_genres_df.head()

# Flatten the 'genre_id' lists into a comma-separated string
title_genres_df['genre_id'] = title_genres_df['genre_id'].apply(lambda x: ','.join(map(str, x)))

# Insert the data into the 'title_genres' table
title_genres_df.to_sql('title_genres', engine, index=False, if_exists='replace')

title_basics_df.head()

# Insert the DataFrame into the database
exploded_genres.to_sql('title_basics', engine, index=False, if_exists='replace')

print(title_basics_df.columns)

# Define a base class for declarative models
Base = declarative_base()


# Define your model class with a primary key and key length
class TitleBasics(Base):
    __tablename__ = 'title_basics'
    tconst = Column(String(20), primary_key=True)  # Here, '20' is the key length
    # Define other columns
    primaryTitle = Column(String)
    startYear = Column(Integer)
    runtimeMinutes =Column(Integer)
    originalTitle =Column(String)
    genres = Column(String)


title_ratings_df.to_sql('title_ratings', engine, index=False, if_exists='replace')


print(title_ratings_df.columns)

# Define the model class for title_ratings
class TitleRatings(Base):
    __tablename__ = 'title_ratings'
    tconst = Column(String(20), primary_key=True)  # Assuming '20' is the key length
    averageRating = Column(Float)
    numVotes = Column(Integer)


tmdb_data_df[['movie_id', 'budget', 'revenue', 'certification']].to_sql('tmdb_data', engine, index=False, if_exists='replace')

print(tmdb_data_df.columns)

# Define your model class with a primary key and key length
class tmdb_data(Base):
    __tablename__ = 'tmdb_data'
    movie_id = Column(String(20), primary_key=True)  # Here, '20' is the key length
    # Define other columns
    budget = Column(Integer)
    revenue = Column(Integer)
    certification =Column(Float)
    

# Define the schema for the 'genres' table
create_genres_table = """
CREATE TABLE IF NOT EXISTS genres (
    tconst VARCHAR(255) PRIMARY KEY,
    genre TEXT
);
"""

# Execute the SQL command to create the table
engine.execute(create_genres_table)


# Execute SQL queries to show the first 5 rows of each table
tables = ['title_basics', 'title_ratings', 'title_genres', 'genres', 'tmdb_data']
for table in tables:
    result = engine.execute(f'SELECT * FROM {table} LIMIT 5')
    print(f'First 5 rows of {table}:')
    for row in result:
        print(row)
    print('\n')


# Show all tables in the 'movies' database
result = engine.execute('SHOW TABLES')
print('Tables in the database:')
for row in result:
    print(row[0])


# Create a session
Session = sessionmaker(bind=engine)
session = Session()

session.commit()
